+++ src/game/util/Prices.java
@@ -4,57 +4,57 @@ public class Prices {
 
 	public static final int NUM_PRICES = 49;
 	
-	public int station_value;
-	public int build_rail;
-	public int build_road;
-	public int build_signals;
-	public int build_bridge;
-	public int build_train_depot;
-	public int build_road_depot;
-	public int build_ship_depot;
-	public int build_tunnel;
-	public int train_station_track;
-	public int train_station_length;
-	public int build_airport;
-	public int build_bus_station;
-	public int build_truck_station;
-	public int build_dock;
-	public int build_railvehicle;
-	public int build_railwagon;
-	public int aircraft_base;
-	public int roadveh_base;
-	public int ship_base;
-	public int build_trees;
-	public int terraform;
-	public int clear_1;
-	public int purchase_land;
-	public int clear_2;
-	public int clear_3;
-	public int remove_trees;
-	public int remove_rail;
-	public int remove_signals;
-	public int clear_bridge;
-	public int remove_train_depot;
-	public int remove_road_depot;
-	public int remove_ship_depot;
-	public int clear_tunnel;
-	public int clear_water;
-	public int remove_rail_station;
-	public int remove_airport;
-	public int remove_bus_station;
-	public int remove_truck_station;
-	public int remove_dock;
-	public int remove_house;
-	public int remove_road;
-	public final int [] running_rail = new int[3];
-	public int aircraft_running;
-	public int roadveh_running;
-	public int ship_running;
-	public int build_industry;
+	public double station_value;
+	public double build_rail;
+	public double build_road;
+	public double build_signals;
+	public double build_bridge;
+	public double build_train_depot;
+	public double build_road_depot;
+	public double build_ship_depot;
+	public double build_tunnel;
+	public double train_station_track;
+	public double train_station_length;
+	public double build_airport;
+	public double build_bus_station;
+	public double build_truck_station;
+	public double build_dock;
+	public double build_railvehicle;
+	public double build_railwagon;
+	public double aircraft_base;
+	public double roadveh_base;
+	public double ship_base;
+	public double build_trees;
+	public double terraform;
+	public double clear_1;
+	public double purchase_land;
+	public double clear_2;
+	public double clear_3;
+	public double remove_trees;
+	public double remove_rail;
+	public double remove_signals;
+	public double clear_bridge;
+	public double remove_train_depot;
+	public double remove_road_depot;
+	public double remove_ship_depot;
+	public double clear_tunnel;
+	public double clear_water;
+	public double remove_rail_station;
+	public double remove_airport;
+	public double remove_bus_station;
+	public double remove_truck_station;
+	public double remove_dock;
+	public double remove_house;
+	public double remove_road;
+	public final double [] running_rail = new double[3];
+	public double aircraft_running;
+	public double roadveh_running;
+	public double ship_running;
+	public double build_industry;
 
 	
 	
-	public int getPrice(int i)
+	public double getPrice(int i)
 	{
 		switch(i)
 		{
@@ -115,7 +115,7 @@ public class Prices {
 	}
 	
 
-	public void setPrice(int i, int price)
+	public void setPrice(int i, double price)
 	{
 		switch(i)
 		{
